+++ DisCatSharp.ApplicationCommands/Workers/ApplicationCommandWorker.cs
